move code into lib/channels deprecate the individual channels
