+++ code/builders/builder-webpack5/templates/virtualModuleModernEntry.js.handlebars
@@ -1,7 +1,7 @@
 import { global } from '@storybook/global';
 
 import { ClientApi, PreviewWeb, addons, composeConfigs } from '@storybook/preview-api';
-import { createChannel as createPostMessageChannel } from '@storybook/channel-postmessage';
+import { createPostMessageChannel } from '@storybook/channels';
 
 import { importFn } from './{{storiesFilename}}';
