+++ code/deprecated/channel-postmessage/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "@storybook/channel-websocket",
+  "name": "@storybook/channel-postmessage",
   "version": "7.1.0-alpha.31",
   "description": "",
   "keywords": [
     "storybook"
   ],
-  "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/channel-websocket",
+  "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/channel-postmessage",
   "bugs": {
     "url": "https://github.com/storybookjs/storybook/issues"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/storybookjs/storybook.git",
-    "directory": "code/lib/channel-websocket"
+    "directory": "code/lib/channel-postmessage"
   },
   "funding": {
     "type": "opencollective",
@@ -22,16 +22,16 @@
   "sideEffects": false,
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "node": "./dist/index.js",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
+      "types": "./dist/entry.d.ts",
+      "node": "./dist/entry.js",
+      "require": "./dist/entry.js",
+      "import": "./dist/entry.mjs"
     },
     "./package.json": "./package.json"
   },
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "main": "dist/entry.js",
+  "module": "dist/entry.mjs",
+  "types": "dist/entry.d.ts",
   "files": [
     "dist/**/*",
     "README.md",
@@ -44,10 +44,7 @@
   },
   "dependencies": {
     "@storybook/channels": "7.1.0-alpha.31",
-    "@storybook/client-logger": "7.1.0-alpha.31",
-    "@storybook/global": "^5.0.0",
-    "telejson": "^7.0.3",
-    "tiny-invariant": "^1.3.1"
+    "@storybook/client-logger": "7.1.0-alpha.31"
   },
   "devDependencies": {
     "typescript": "~4.9.3"
@@ -57,8 +54,9 @@
   },
   "bundler": {
     "entries": [
-      "./src/index.ts"
-    ]
+      "./src/entry.ts"
+    ],
+    "shim": "@storybook/channels/dist/postmessage/index"
   },
   "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16"
 }
