+++ code/deprecated/channel-websocket/jest.config.js
