+++ code/deprecated/channel-websocket/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "@storybook/channel-postmessage",
+  "name": "@storybook/channel-websocket",
   "version": "7.1.0-alpha.31",
   "description": "",
   "keywords": [
     "storybook"
   ],
-  "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/channel-postmessage",
+  "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/channel-websocket",
   "bugs": {
     "url": "https://github.com/storybookjs/storybook/issues"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/storybookjs/storybook.git",
-    "directory": "code/lib/channel-postmessage"
+    "directory": "code/lib/channel-websocket"
   },
   "funding": {
     "type": "opencollective",
@@ -22,16 +22,16 @@
   "sideEffects": false,
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "node": "./dist/index.js",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
+      "types": "./dist/entry.d.ts",
+      "node": "./dist/entry.js",
+      "require": "./dist/entry.js",
+      "import": "./dist/entry.mjs"
     },
     "./package.json": "./package.json"
   },
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "main": "dist/entry.js",
+  "module": "dist/entry.mjs",
+  "types": "dist/entry.d.ts",
   "files": [
     "dist/**/*",
     "README.md",
@@ -40,17 +40,11 @@
   ],
   "scripts": {
     "check": "../../../scripts/node_modules/.bin/tsc --noEmit",
-    "prep": "../../../scripts/prepare/bundle.ts"
+    "prep": "../../../scripts/prepare/facade.ts"
   },
   "dependencies": {
-    "@storybook/channel-websocket": "7.1.0-alpha.31",
     "@storybook/channels": "7.1.0-alpha.31",
-    "@storybook/client-logger": "7.1.0-alpha.31",
-    "@storybook/core-events": "7.1.0-alpha.31",
-    "@storybook/global": "^5.0.0",
-    "qs": "^6.10.0",
-    "telejson": "^7.0.3",
-    "tiny-invariant": "^1.3.1"
+    "@storybook/client-logger": "7.1.0-alpha.31"
   },
   "devDependencies": {
     "typescript": "~4.9.3"
@@ -60,8 +54,9 @@
   },
   "bundler": {
     "entries": [
-      "./src/index.ts"
-    ]
+      "./src/entry.ts"
+    ],
+    "shim": "@storybook/channels/dist/websocket/index"
   },
   "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16"
 }
