+++ code/deprecated/channel-websocket/tsconfig.json
