--- code/lib/channel-websocket/src/typings.d.ts
@@ -1,2 +1,0 @@
-declare module 'json-fn';
-declare var CONFIG_TYPE: 'DEVELOPMENT' | 'PRODUCTION';
