+++ code/lib/channels/package.json
@@ -27,6 +27,18 @@
       "require": "./dist/index.js",
       "import": "./dist/index.mjs"
     },
+    "./dist/postmessage/index": {
+      "types": "./dist/postmessage/index.d.ts",
+      "node": "./dist/postmessage/index.js",
+      "require": "./dist/postmessage/index.js",
+      "import": "./dist/postmessage/index.mjs"
+    },
+    "./dist/websocket/index": {
+      "types": "./dist/websocket/index.d.ts",
+      "node": "./dist/websocket/index.js",
+      "require": "./dist/websocket/index.js",
+      "import": "./dist/websocket/index.mjs"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -42,6 +54,15 @@
     "check": "../../../scripts/node_modules/.bin/tsc --noEmit",
     "prep": "../../../scripts/prepare/bundle.ts"
   },
+  "dependencies": {
+    "@storybook/channels": "7.1.0-alpha.31",
+    "@storybook/client-logger": "7.1.0-alpha.31",
+    "@storybook/core-events": "7.1.0-alpha.31",
+    "@storybook/global": "^5.0.0",
+    "qs": "^6.10.0",
+    "telejson": "^7.0.3",
+    "tiny-invariant": "^1.3.1"
+  },
   "devDependencies": {
     "typescript": "~4.9.3"
   },
@@ -50,7 +71,9 @@
   },
   "bundler": {
     "entries": [
-      "./src/index.ts"
+      "./src/index.ts",
+      "./src/postmessage/index.ts",
+      "./src/websocket/index.ts"
     ]
   },
   "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16"
