+++ code/lib/channels/src/typings.d.ts
