+++ code/lib/preview-api/src/modules/core-client/start.test.ts
@@ -37,7 +37,7 @@ jest.mock('@storybook/global', () => ({
 
 // console.log(global);
 
-jest.mock('@storybook/channel-postmessage', () => ({ createChannel: () => mockChannel }));
+jest.mock('@storybook/channels', () => ({ createPostmessageChannel: () => mockChannel }));
 jest.mock('@storybook/client-logger');
 jest.mock('react-dom');
