+++ code/lib/preview-api/src/modules/preview-web/PreviewWeb.integration.test.ts
@@ -24,7 +24,7 @@ import {
 //   - ie. from`renderToCanvas()` (stories) or`ReactDOM.render()` (docs) in.
 // This file lets them rip.
 
-jest.mock('@storybook/channel-postmessage', () => ({ createChannel: () => mockChannel }));
+jest.mock('@storybook/channels', () => ({ createPostMessageChannel: () => mockChannel }));
 jest.mock('@storybook/client-logger');
 
 jest.mock('./WebView');
